Factor Version out to pep_440.py
