+++ pex/cli/commands/lockfile/json_codec.py
@@ -9,6 +9,7 @@ from pex import compatibility
 from pex.cli.commands.lockfile import ParseError
 from pex.cli.commands.lockfile.lockfile import Lockfile
 from pex.enum import Enum
+from pex.pep_440 import Version
 from pex.pep_503 import ProjectName
 from pex.resolve.locked_resolve import (
     Artifact,
@@ -17,7 +18,6 @@ from pex.resolve.locked_resolve import (
     LockedResolve,
     LockStyle,
     Pin,
-    Version,
 )
 from pex.resolve.resolver_configuration import ResolverVersion
 from pex.third_party.packaging import tags
